Refactor help dialog management into a dedicated controller and remove legacy code
