+++ web/index.html
@@ -119,7 +119,7 @@
   <!-- =========================
        HELP DIALOG (native <dialog> with ::backdrop)
        ========================= -->
-  <dialog id="help-dialog" class="help-dialog" aria-label="Help">
+  <dialog id="help-dialog" class="help-dialog" closedby="any" aria-label="Help">
 
     <!-- Dark header -->
     <div class="bg-panel-toolbar px-6 py-4 flex items-center justify-between select-none rounded-t-panel">
